more fixes to shell script

+++ deploy-local.sh
@@ -3,5 +3,5 @@
 
 echo delpoying...
 
-rm -R /var/www/html/wp-content/themes/wp-andreas225
+[ -d /var/www/html/wp-content/themes/wp-andreas225] && rm -R /var/www/html/wp-content/themes/wp-andreas225
 mv ./wp-andreas225 /fifteensquared-theme$ ls /var/www/html/wp-content/themes
